♻️ Move all combinators to new submodules under the `combinators` module
That way they have their own file along their tests. Refactor made with the help of Claude Sonnet 4.6.
