--- src/tests/maybe.rs
@@ -1,66 +1,0 @@
-use crate::*;
-use input::position::Position;
-
-#[test]
-fn empty() {
-	let parser = is('h');
-	let mut input = Input::new_from_chars("".chars(), None);
-	let parser_maybe = maybe(&parser);
-	assert_eq!(parser_maybe(&mut input), Ok(None));
-	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
-}
-
-#[test]
-fn empty_shortcut() {
-	let parser = is('h').maybe();
-	let mut input = Input::new_from_chars("".chars(), None);
-	assert_eq!(parser(&mut input), Ok(None));
-	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
-}
-
-#[test]
-fn success() {
-	let parser = is('h');
-	let mut input = Input::new_from_chars("h".chars(), None);
-	let parser_maybe = maybe(&parser);
-	assert_eq!(parser_maybe(&mut input), Ok(Some('h')));
-	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
-}
-
-#[test]
-fn success_shortcut() {
-	let parser = is('h').maybe();
-	let mut input = Input::new_from_chars("h".chars(), None);
-	assert_eq!(parser(&mut input), Ok(Some('h')));
-	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
-}
-
-#[test]
-fn fail_non_consuming() {
-	let parser = is('h');
-	let mut input = Input::new_from_chars("j".chars(), None);
-	let parser_maybe = maybe(&parser);
-	assert_eq!(parser_maybe(&mut input), Ok(None));
-	assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
-}
-
-#[test]
-fn fail_consuming() {
-	let parser = |input: &mut StringInput| match any()(input) {
-		Ok(token) => {
-			if token == 'h' {
-				Ok(1)
-			} else {
-				Err(Error::UnexpectedToken(None, Position::new(1, 1)))
-			}
-		}
-		Err(e) => Err(e),
-	};
-	let mut input = Input::new_from_chars("j".chars(), None);
-	let parser_maybe = maybe(&parser);
-	assert_eq!(
-		parser_maybe(&mut input),
-		Err(Error::UnexpectedToken(None, Position::new(1, 1)))
-	);
-	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
-}
