--- src/tests/mod.rs
@@ -1,22 +1,0 @@
-mod and;
-mod and_then;
-mod any;
-mod attempt;
-mod between;
-mod chain_left;
-mod chain_right;
-mod choice;
-mod count;
-mod end_of_input;
-mod is;
-mod look_ahead;
-mod many0;
-mod many1;
-mod many_until;
-mod map;
-mod maybe;
-mod not_followed_by;
-mod option;
-mod satisfy;
-mod separated_by0;
-mod separated_by1;
